Stated working on handlers
Handlers are the individual components of the web server. These can be configured to have their own settings.

The first (TBA) handler to be available is the file-system handler, which will sent files to the clients when requested.

Handlers have their own root, so a server can be made up of multiple types of handlers.
This enables the possibility of having CGI.