Added global settings
Global setting is a header file useful for storing things that different parts of the program might need, which results in consistency between componenents. The first value to exist in the global settings file is the value of the 'Server' response header, which can now be easily changed in the configuration file.