Fixed a bug caused by implementing headers_ready callbacks
Some headers weren't added before callback_info->header_ready(...) so these were discarded by the header writer. I didn't notice this at first because I was working with HTTP/2 which has a standard DATA-frame size which Chromium apparantly uses as an alternative for content-length when that header isn't in the list.