Fixed a bug in the io_read_until function
The function previously returned 0 on a I/O error, but because the position could also be 0, the function return value was ambiguous. This resulted in headers with a key (a.k.a. name) of length 2 (e.g. 'TE') was seen as an error, because the CRLF is also two characters long. This is now fixed! :)